ci: create initial gitlab-ci jobs
Adds two simple jobs for building the library for Debian 10 and CentOS 8

They will pull down docker images containing the dependencies needed for
libpEpAdapter, then proceed with the build. Upon successful compilation,
an image containing the ouput (and dependencies) will be pushed to a
docker registry on the CI infrastructure.

Note about the docker images in use:
Currently they pull the latest docker images containing pEpEngine,
Sequoia, and other dependencies. We want to specify the exact required
version in future updates to the jobs.