CORE-205 Clear sync event deque before starting sync protocol
This is done to cooperate with the apps.
After user leaves device group one possible workaround to be able to sync again is stopping sync, which adds a null event to the deque.
Then when user tries to start sync again we don't want to find that null event which will cause sync to stop again.